+++ lib/core/constants/strings.dart
@@ -2,6 +2,6 @@ class Strings {
   // App
   static const appTitle = 'Swickky';
   static const appDesc = 'Swiggy clone';
-  static const appVersion = 'Version: 1.0.3';
+  static const appVersion = 'Version: 1.0.4';
   static const appDev = 'Developed by Chetan R Billava';
 }